Resolve some leftovers

+++ conseil-lorre/src/main/scala/tech/cryptonomic/conseil/indexer/logging/LorreProgressLogging.scala
@@ -16,15 +16,15 @@ trait LorreProgressLogging {
     * @param processStartNanos a nano-time from jvm monotonic time, used to identify when the whole processing operation began
     * @param processed how many entities were processed at the current checkpoint
     */
-  def logProcessingProgress(entityName: String, totalToProcess: BigDecimal, processStartNanos: Long)(
+  def logProcessingProgress(entityName: String, totalToProcess: Long, processStartNanos: Long)(
       processed: Int
   ): Unit = {
     val elapsed = System.nanoTime() - processStartNanos
-    val progress = BigDecimal(processed) / totalToProcess
+    val progress = processed.toDouble / totalToProcess
     logger.info("================================== Progress Report ==================================")
     logger.info("Completed processing {}% of total requested {}s", "%.2f".format(progress * 100), entityName)
 
-    val etaMins = Duration(scala.math.ceil(elapsed / progress.doubleValue()) - elapsed, NANOSECONDS).toMinutes
+    val etaMins = Duration(scala.math.ceil(elapsed / progress) - elapsed, NANOSECONDS).toMinutes
     if (processed < totalToProcess && etaMins > 1) logger.info("Estimated time to finish is around {} minutes", etaMins)
     logger.info("=====================================================================================")
   }
